Tidy up modal/dropdown: remove fieldDefinition and summaryFields
