+++ packages/dataviews/src/components/dataform-layouts/panel/utils/get-field-definition-and-summary-fields.ts
@@ -1,0 +1,70 @@
+import type {
+	NormalizedFormField,
+	NormalizedField,
+	NormalizedPanelLayout,
+} from '../../../../types';
+import { getSummaryFields } from '../../get-summary-fields';
+
+const getFieldDefinition = < Item >(
+	field: NormalizedFormField,
+	fields: NormalizedField< Item >[]
+) => {
+	const fieldDefinition = fields.find( ( _field ) => _field.id === field.id );
+
+	if ( ! fieldDefinition ) {
+		return fields.find( ( _field ) => {
+			if ( !! field.children ) {
+				const simpleChildren = field.children.filter(
+					( child ) => ! child.children
+				);
+
+				if ( simpleChildren.length === 0 ) {
+					return false;
+				}
+
+				return _field.id === simpleChildren[ 0 ].id;
+			}
+
+			return _field.id === field.id;
+		} );
+	}
+
+	return fieldDefinition;
+};
+
+/**
+ * Determines the field definition and summary fields for a panel layout.
+ *
+ * Summary fields are determined with the following priority:
+ * 1. Use layout.summary fields if they exist
+ * 2. Fall back to the field definition that matches the form field's id
+ * 3. If the form field id doesn't exist, pick the first child field
+ * 4. If no field definition is found, return empty summary fields
+ *
+ * @param layout - The normalized panel layout configuration
+ * @param field  - The form field to get definition for
+ * @param fields - Array of normalized field definitions
+ * @return Object containing fieldDefinition and summaryFields
+ */
+const getFieldDefinitionAndSummaryFields = < Item >(
+	layout: NormalizedPanelLayout,
+	field: NormalizedFormField,
+	fields: NormalizedField< Item >[]
+) => {
+	const summaryFields = getSummaryFields( layout.summary, fields );
+	const fieldDefinition = getFieldDefinition( field, fields );
+
+	if ( summaryFields.length === 0 ) {
+		return {
+			summaryFields: fieldDefinition ? [ fieldDefinition ] : [],
+			fieldDefinition,
+		};
+	}
+
+	return {
+		summaryFields,
+		fieldDefinition,
+	};
+};
+
+export default getFieldDefinitionAndSummaryFields;
